Expose more information about USB devices
Makes `adapters::usb::HsmDevice` wrap `libusb::Device` and stores it
with the `UsbAdapter`, and exposes additional info like bus and device
addresses.
