+++ src/adapters/http/macros.rs
@@ -1,0 +1,9 @@
+/// Write consistent `debug!(...) lines for adapters
+macro_rules! http_debug {
+    ($adapter:expr, $msg:expr) => {
+        debug!("yubihsm-connector({}) {}", $adapter.host, $msg);
+    };
+    ($adapter:expr, $fmt:expr, $($arg:tt)+) => {
+        debug!(concat!("yubihsm-connector({}) ", $fmt), $adapter.host, $($arg)+);
+    };
+}
